refactor(search): split form coordination
